refactor(trace-viewer): extract SpawnTraceViewer and TraceViewerFactory
