+++ src/testModel.ts
@@ -30,8 +30,7 @@ import type { PlaywrightTestRunOptions, RunHooks, TestConfig } from './playwrigh
 import { PlaywrightTestCLI } from './playwrightTestCLI';
 import { upstreamTreeItem } from './testTree';
 import { collectTestIds } from './upstream/testTree';
-import { SpawnTraceViewer, TraceViewer } from './traceViewer';
-import { EmbeddedTraceViewer } from './embeddedTraceViewer';
+import { TraceViewer, TraceViewerFactory } from './traceViewer';
 
 export type TestEntry = reporterTypes.TestCase | reporterTypes.Suite;
 
@@ -57,64 +56,6 @@ export type TestModelEmbedder = {
 type Watch = {
   include: readonly vscodeTypes.TestItem[] | undefined;
 };
-
-class TraceViewerFactory {
-  private _embedder: TestModelEmbedder;
-  private _vscode: vscodeTypes.VSCode;
-  private _config: TestConfig;
-  private _testServer?: PlaywrightTestServer;
-
-  constructor(collection: TestModelCollection, config: TestConfig, testServer?: PlaywrightTestServer) {
-    this._vscode = collection.vscode;
-    this._embedder = collection.embedder;
-    this._config = config;
-    this._testServer = testServer;
-  }
-
-  create(): TraceViewer | null {
-    if (this._checkEmbeddedSupport(true))
-      return new EmbeddedTraceViewer(this._vscode, this._embedder.context.extensionUri, this._config, this._testServer!);
-    else if (this._checkSpawnSupport(true))
-      return new SpawnTraceViewer(this._vscode, this._embedder.envProvider, this._config);
-    return null;
-  }
-
-  isSupported(traceViewer: TraceViewer) {
-    if (traceViewer instanceof EmbeddedTraceViewer && this._checkEmbeddedSupport())
-      return true;
-    if (traceViewer instanceof SpawnTraceViewer && this._checkSpawnSupport())
-      return true;
-    return false;
-  }
-
-  private _checkSpawnSupport(userGesture?: boolean) {
-    if (!this._embedder.settingsModel.showTrace.get())
-      return false;
-    if (!this._checkVersion(1.35, this._vscode.l10n.t('this feature'), userGesture))
-      return false;
-    return true;
-  }
-
-  private _checkEmbeddedSupport(userGesture?: boolean) {
-    if (!this._embedder.settingsModel.showTrace.get() || !this._embedder.settingsModel.embeddedTraceViewer.get())
-      return false;
-    if (!this._checkVersion(1.46, this._vscode.l10n.t('embedded trace viewer'), userGesture))
-      return false;
-    return true;
-  }
-
-  private _checkVersion(version: number, message: string, userGesture?: boolean) {
-    if (this._config.version < version) {
-      if (userGesture) {
-        this._vscode.window.showWarningMessage(
-            this._vscode.l10n.t('Playwright v{0}+ is required for {1} to work, v{2} found', version, message, this._config.version)
-        );
-      }
-      return false;
-    }
-    return true;
-  }
-}
 
 export class TestModel extends DisposableBase {
   private _vscode: vscodeTypes.VSCode;
